+++ AGENTS.md
@@ -1,7 +1,7 @@
 <!-- gitnexus:start -->
 # GitNexus — Code Intelligence
 
-This project is indexed by GitNexus as **lyra** (16291 symbols, 37698 relationships, 300 execution flows). Use the GitNexus MCP tools to understand code, assess impact, and navigate safely.
+This project is indexed by GitNexus as **753-eliminate-all-300-line-file-exemptions** (16940 symbols, 39176 relationships, 300 execution flows). Use the GitNexus MCP tools to understand code, assess impact, and navigate safely.
 
 > If any GitNexus tool warns the index is stale, run `npx gitnexus analyze` in terminal first.
 
@@ -17,7 +17,7 @@ This project is indexed by GitNexus as **lyra** (16291 symbols, 37698 relationsh
 
 1. `gitnexus_query({query: "<error or symptom>"})` — find execution flows related to the issue
 2. `gitnexus_context({name: "<suspect function>"})` — see all callers, callees, and process participation
-3. `READ gitnexus://repo/lyra/process/{processName}` — trace the full execution flow step by step
+3. `READ gitnexus://repo/753-eliminate-all-300-line-file-exemptions/process/{processName}` — trace the full execution flow step by step
 4. For regressions: `gitnexus_detect_changes({scope: "compare", base_ref: "main"})` — see what your branch changed
 
 ## When Refactoring
@@ -56,10 +56,10 @@ This project is indexed by GitNexus as **lyra** (16291 symbols, 37698 relationsh
 
 | Resource | Use for |
 |----------|---------|
-| `gitnexus://repo/lyra/context` | Codebase overview, check index freshness |
-| `gitnexus://repo/lyra/clusters` | All functional areas |
-| `gitnexus://repo/lyra/processes` | All execution flows |
-| `gitnexus://repo/lyra/process/{name}` | Step-by-step execution trace |
+| `gitnexus://repo/753-eliminate-all-300-line-file-exemptions/context` | Codebase overview, check index freshness |
+| `gitnexus://repo/753-eliminate-all-300-line-file-exemptions/clusters` | All functional areas |
+| `gitnexus://repo/753-eliminate-all-300-line-file-exemptions/processes` | All execution flows |
+| `gitnexus://repo/753-eliminate-all-300-line-file-exemptions/process/{name}` | Step-by-step execution trace |
 
 ## Self-Check Before Finishing
